requirements.txt
streamlit
google-generativeai
easyocr
Pillow
numpy
python-docx
PyPDF2